add functionality for multi doc type indexes
lossen validation so that indexes can be built via inheritance of
document classes.  A parent class can define the general index strategy
and name while children can define doc types and mappings.  Then
querying can be done on the parent or any of the individual children.

The key to this is the subclasses config variable. This allows a
elasticity document parent class to know about it's children who are
being used to make up it's mappings and doc types.