Es 2 (#41)
* bump to support any elasticsearch  version greater than 1.0

* change dependency to allow for multiple versions of elasticsearch

elasticsearch-ruby separates version dependencies from elasticsearch by
major version bumps, but we do not utilize a hard dependency on any of
the changes in the version bumps, so we want to loosen the dependency
requirement so users can use the version of elasticsearch that works for
them.  The main change is the removal of scan api which we intuit from gem version.