update remap method to remove undefined mapping properties
If a field is not explicitly defined in an index's mapping we should not
persist it via the documents we copy over as part of remap.  This allows
for a way to get around the limitation in lucene where you can't delete
fields from index mappings without recreating the index from scratch.