allow msearch to return errors rather than raising
to more correctly mimic the _msearch API,  allow msearch
to return the errors about an index rather than raising
an exception.  This will let the consumers of the API
display results from the other indexes of the msearch
and handle errors on their side.

Also return a 'null' result set along with the error
and exception info so the consumer can more gracefully
handle the errors without additional case discrimination