retry deleting the index after certain exceptions during remap
If a snapshot is in progress when we try to delete the old index
during a remap, the whole remap will fail. And it is only due to
bad timing.
That is an expensive errors since some remap can take 10s of hours.
So lets give the option of retrying that delete so we don't have
to waste all the cycles used during the remap