replace deprecated ActiveSupport::ProxyObject with BasicObject
The use of `ActiveSupport::ProxyObject` was deprecated via [1] and
was fully removed in Rails 8.0. This updates the code to switch from
ProxyObject to Ruby's built-in `BasicObject` [2].

The `BasicObject` has been available in Ruby since v1.9, and given
that this gem has direct dependencies on activemodel/activesupport/
activerecord >= 5.2.0 which themselves require a minimum Ruby version
of 2.2 [3], this shouldn't introduce breaking changes for the range of
Ruby versions currently allowed by this gem. Additionally, the `raise`
method that was defined by `ProxyObject` [4] was not being leveraged
in the gem's code, so this should be a clean replacement, however,
we can define `raise` for these classes if we want to be extra
conservative with the change.

[1] https://github.com/rails/rails/pull/51638
[2] https://docs.ruby-lang.org/en/3.4/BasicObject.html
[3] https://rubygems.org/gems/activemodel/versions/5.2.0
[4] https://github.com/rails/rails/blob/6f0d1ad14b92b9f5906e44740fce8b4f1c7075dc/activesupport/lib/active_support/proxy_object.rb#L13-L15